requirements.txt
beautifulsoup4==4.14.3
python-dotenv==1.2.1
